Create Web control LED
Introduction
The reason why Raspberry Pi is so amazing is the networking function and support by
numerous open source web applications. With it, your own Internet of Things system can be
constructed conveniently. In this lesson, you will control an LED based on the web with
Raspberry Pi. You can use your mobile phone, tablet pc or computer to control the LED too, as
long as it is in the same Local Area Network (LAN) with your Raspberry Pi.
Components
- 1* Raspberry Pi
- 1* Breadboard
- 1* LED
- 1* Resistor (220Ω)
- Jumper wires